requirements.txt
ipykernel
jupyter
jupyterlab
pandas
numpy
matplotlib
seaborn
statsmodels
tbb==2019.0
umap-learn==0.4.2
scikit-learn==0.22.2
